==> dataAnalyses/1.batchCorrection_genes.Rmd <==
---
title: "Batch correction for gene count table"
author:
- tabbassidaloii
output:
 md_document:
    variant: markdown_github
always_allow_html: yes
editor_options:
  chunk_output_type: console
---

'Note:' The muscle abbreviations were GRA (for Gracilis), SED (semitendinosus distal),  SEM (semitendinosus middle)), GAL (gastrocnemius lateralis), REF (rectus femoris), VAM (vastus medialis) and VAL (vastus lateralis) in all the analyses (scripts) and they were only changed to GR, STM, STD, GL, RF, VM and VL for the plots (shown in the publication).

```{r setup, include = FALSE}
#Empty the R environment
rm(list = ls())

#Set your working environment to the location where your current source file is saved into.
setwd(dirname(rstudioapi::getSourceEditorContext()$path))

#Check if libraries are already installed, otherwise install them
if(!"dplyr" %in% installed.packages()) install.packages("dplyr")
if(!"tibble" %in% installed.packages()) install.packages("tibble")
if(!"ggplot2" %in% installed.packages()) install.packages("ggplot2")
if(!"scales" %in% installed.packages()) install.packages("scales")
if(!require("BiocManager", quietly = TRUE)) install.packages("BiocManager")
if(!"limma" %in% installed.packages()) BiocManager::install("limma")
if(!"edgeR" %in% installed.packages()) BiocManager::install("edgeR")
if(!"biomaRt" %in% installed.packages()) BiocManager::install("biomaRt")
if(!"knitr" %in% installed.packages()) install.packages("knitr")
if(!"factoextra" %in% installed.packages()) install.packages("factoextra")
if(!"gplots" %in% installed.packages()) install.packages("gplots")
if(!"RColorBrewer" %in% installed.packages()) install.packages("RColorBrewer")
if(!"ggpubr" %in% installed.packages()) install.packages("ggpubr")
if(!"reshape2" %in% installed.packages()) install.packages("reshape2")

#Load installed libraries
suppressPackageStartupMessages({
  library(dplyr)
  library(ggplot2)
  library(scales)
  library(limma)
  library(edgeR)
  library(biomaRt)
  library(knitr)
  library(factoextra)
  library(gplots)
  library(RColorBrewer)
  library(ggpubr)
})

#RNA isolation protocol for samples isolated with both protocols
withKit <- c("MD06_GRA_Sam25", "MD07_GRA_Sam23", "MD08_GRA_Sam65", 
              "MD10_GRA_Sam106", "MD12_GRA_Sam139")
withoutKit <- c("MD06_GRA_Sam148", "MD07_GRA_Sam73_RedoSam15", "MD08_GRA_Sam152_RedoSam39", 
                 "MD10_GRA_Sam36", "MD12_GRA_Sam116")
#Set colors
col.muscle <- c("GAL" = "#999999", "GRA" = "#0099FF", "REF" = "#FF9999", 
                 "SED" = "#0033FF", "SEM" = "#0066FF", "VAL" = "#FF6666", 
                 "VAM" = "#FF3333")
```

Comparing the DE results of(A)all batches and(B)only 4 batches, we learned that the including the batch as a covariate in the linear mixed models will not correct for the batch. So we used the ComBat-Seq method to correct for batch.

Why including in the mixed model could not correct for batch:
If the only difference between batches would be the mean expression level of gene, the batch in the LMM would correct for batch but genes have different dispersion across the batches. 

# Batch effect correction for genes
In ComBat-Seq, we can specify biological covariates, whose signals will be preserved in the adjusted data.
- One biological variable: group parameter (we include muscle as a biological covariates to preserve its effect)

```{r ComBatSeq, echo = FALSE, warning = FALSE, message = FALSE}
#Load the samples passed the QC step
passedQC <- unlist(read.table("inputs/sampleListPassedQualityControl.txt", as.is = TRUE), use.names = FALSE)

#Read the count-data
HTseqCount <- read.csv("inputs/countTables/primarySequencingFragmentsPerGene", sep = "\t")
reSeqHTseqCount <- read.csv("inputs/countTables/reSequencingFragmentsPerGene", sep = "\t")

HTseqCount <- merge(HTseqCount, reSeqHTseqCount) %>% 
  filter(grepl("ENSG", feature)) %>% #remove extra rows
  dplyr::select(one_of(c("feature", passedQC))) %>% #filtering samples
  filter_at(vars(-feature), any_vars(. !=  0)) %>% #removing genes with zero in all the rows
  tibble::column_to_rownames("feature") %>%
  rename(MD10_REF_Sam113_RedoSam25 = MD08_REF_Sam113_RedoSam25)#sample swap check fixation

keep.sam <- colnames(HTseqCount)[! colnames(HTseqCount) %in%  withKit]#removing samples sequenced to compare two RNA isolation protocols
HTseqCount <- HTseqCount[, keep.sam]

SampleData <- data.frame(sample = colnames(HTseqCount), 
                         sampleID = as.factor(gsub(".*_", "", colnames(HTseqCount))), 
                         individual = as.factor(gsub("_.*", "", colnames(HTseqCount))), 
                         muscle = as.factor(gsub(".*_", "", gsub("_Sam.*", "", colnames(HTseqCount)))), 
                         batch = as.factor(
                           ifelse(gsub(".*_", "", colnames(HTseqCount)) %in% paste0("Sam", 1:39), "1", 
                                  ifelse(gsub(".*_", "", colnames(HTseqCount)) %in% paste0("Sam", 40:78), "2", 
                                         ifelse(gsub(".*_", "", colnames(HTseqCount)) %in% paste0("Sam", 79:117), "3", 
                                                ifelse(gsub(".*_", "", colnames(HTseqCount)) %in% paste0("Sam", 118:156), 
                                                       "4", "5"))))))

if (all(colnames(HTseqCount) == SampleData$sample))
  countAdjBatchMus <- sva::ComBat_seq(as.matrix(HTseqCount), batch = SampleData$batch, group = SampleData$muscle)
rm (list = setdiff(ls(), c("SampleData", "HTseqCount", "countAdjBatchMus", "withKit", "withoutKit", "col.muscle")))
```

### Library size for each samples
```{r libSizePlot1, echo = FALSE, warning = FALSE, message = FALSE, fig.height = 6, fig.width = 10}
Size <- data.frame(Raw = colSums(HTseqCount), 
                    AdjBatchMus = colSums(countAdjBatchMus)) %>%
  tibble::rownames_to_column() %>% mutate(rowname = gsub("_Sam.*", "", rowname)) %>% reshape2::melt()
(SizeAll <- ggplot(Size, 
                   aes(x = rowname, y = value))+
  geom_point(aes(colour = factor(variable)))+
  scale_y_continuous(labels = comma)+
  ggtitle("#Assigned reads in all the samples")+
  labs(x = element_blank(), y = element_blank(), col = "Data")+
  theme(axis.text.x = element_text(angle = 90, hjust =  1, vjust = 0.5)))
```

## Generating DGEList-object
```{r DGEList, echo = FALSE, warning = FALSE, message = FALSE}
D_Raw <- DGEList(counts = HTseqCount)
D_BatchMus <- DGEList(counts = countAdjBatchMus)

#Adding sample information
if (all(colnames(D_Raw) == SampleData$sample)){
  D_Raw$samples$sampleID <- SampleData$sampleID
  D_Raw$samples$individual <- SampleData$individual
  D_Raw$samples$group <- SampleData$muscle
  D_Raw$samples$batch <- SampleData$batch
  D_Raw$samples$lane <- as.factor(ifelse(gsub(".*_", "", colnames(D_Raw)) %in% paste0("Sam", 1:78), "L003.1", 
                                         ifelse(gsub(".*_", "", colnames(D_Raw)) %in% paste0("Sam", 79:156), "L004.1", 
                                                ifelse(gsub(".*_", "", colnames(D_Raw)) %in% 
                                                         paste0("RedoSam", 1:24), "L002.2", "L004.2"))))
  D_Raw$samples$IsolationRNA <- ifelse(D_Raw$samples$individual %in% 
                                         paste0("MD", c(11, 13:31))| rownames(D_Raw$samples) %in% 
                                         withKit, "withKit", "withoutKit")
  D_Raw$samples$Hospital <- ifelse(D_Raw$samples$individual %in% 
                                     paste0("MD", c(26:31)), "EMC", "HMC")
  QCLab <- read.csv("inputs/labReportForQualityControlCheck.csv", sep = ",", as.is = T)
  D_Raw$samples$RIN <- QCLab$RIN[match(gsub("_Re.*", "", rownames(D_Raw$samples)), QCLab$Sample)]
  D_Raw$samples$RIN[rownames(D_Raw$samples) ==  "MD10_REF_Sam113_RedoSam25"]<- QCLab$RIN[QCLab$Sample == "MD08_REF_Sam113"]
  D_Raw$samples$RIN[D_Raw$samples$RIN == 0] = NA
  D_Raw$samples$Concen. <- QCLab$Concen.[match(gsub("_Re.*", "", rownames(D_Raw$samples)), QCLab$Sample)]
  D_Raw$samples$Concen.[rownames(D_Raw$samples)==  "MD10_REF_Sam113_RedoSam25"]<- QCLab$Concen.[QCLab$Sample == "MD08_REF_Sam113"]
  AgeData <- data.frame(individual = c("MD06", "MD07", "MD08", "MD10", "MD11", "MD12", 
                                       "MD13", "MD14", "MD15", "MD16", "MD17", "MD18", 
                                       "MD19", "MD20", "MD21", "MD22", "MD26", "MD27", 
                                       "MD28", "MD31"), 
                        age = c(19, 20, 22, 29, 24, 32, 
                                25, 26, 24, 21, 25, 28, 
                                31, 25, 28, 26, 22, 28, 
                                20, 24))
  D_Raw$samples$age <- AgeData$age[match(D_Raw$samples$individual, AgeData$individual)]
}

if(all(colnames(D_Raw)== colnames(D_BatchMus))) 
  D_BatchMus$samples <- D_Raw$samples

rm(list = setdiff(ls(), c("HTseqCount", "countAdjBatchMus", "D_Raw", "D_BatchMus", "withKit", "withoutKit", "col.muscle")))
```

### Gene annotations
Using biomart and considering the same ensembl version used for the aligning the reads
- Ensembl Archive Release 98(September 2019)

```{r geneAnnotations, echo = FALSE, warning = FALSE, message = FALSE}
#Add gene annotations
ensembl <- useEnsembl(biomart = "ensembl", dataset = "hsapiens_gene_ensembl", host =  "http://sep2019.archive.ensembl.org/")
GeneName <- getBM(attributes = c('ensembl_gene_id', 'hgnc_symbol', 'chromosome_name', 'gene_biotype'), mart = ensembl)
GeneIDs <- GeneName %>% filter(ensembl_gene_id %in% rownames(D_Raw))
```

```{r duplicatedGeneSymbols, echo = FALSE, warning = FALSE}
#Duplicated gene symbols 
DuplicatedGeneSymbols <- GeneIDs %>% filter(hgnc_symbol !=  "") %>% filter(hgnc_symbol %in% names(table(GeneIDs$hgnc_symbol)[table(GeneIDs$hgnc_symbol)!=  1])) %>% arrange(hgnc_symbol)
print(kable(DuplicatedGeneSymbols, align = "c"))
```

#### Duplicated ensembl gene IDs 
```{r duplicatedENSGIDs, echo = FALSE, warning = FALSE}
#Duplicated Ensembl gene IDs 
DuplicatedENSIDs <- GeneIDs %>% filter(ensembl_gene_id %in% names(table(GeneIDs$ensembl_gene_id)[table(GeneIDs$ensembl_gene_id)!=  1]))
kable(DuplicatedENSIDs, align = "c")

GeneIDs <- GeneIDs[! duplicated(GeneIDs$ensembl_gene_id), ]
GeneIDs <- GeneIDs[order(match(GeneIDs$ensembl_gene_id, rownames(D_Raw))), ]
if(all(rownames(D_Raw)== rownames(D_BatchMus)))
  D_Raw$genes <- D_BatchMus$genes <- GeneIDs

rm (list = setdiff(ls(), c("HTseqCount", "countAdjBatchMus", "D_Raw", "D_BatchMus", "withKit", "withoutKit", "col.muscle")))
```

## Transforming and removing low expressed genes
### Raw data
```{r transFrilteringPlot1, echo = FALSE, warning = FALSE, fig.height = 6, fig.width = 10,, fig.show = 'hold'}
#Transformations
cpm <- cpm(D_Raw)
lcpm <- cpm(D_Raw, log = TRUE)
#Density plot
nsamples <- ncol(D_Raw)

par(mfrow = c(1, 2))
plot(density(lcpm[, 1]), ylim = c(0, 1.35), col = col.muscle[grep(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm))[1], names(col.muscle))], lwd = 2, las = 2, main = "Raw Data", xlab = "logCPM")
for(i in 2 : nsamples){
  den <- density(lcpm[, i])
  lines(den$x, den$y, lwd = 2, col = col.muscle[grep(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm))[i], names(col.muscle))])
}

#Remove genes that are lowly expressed
keep.exprs <- filterByExpr(D_Raw, group = D_Raw$samples$group, min.count = 10)
D_Raw <- D_Raw[keep.exprs,, keep.lib.sizes = FALSE]

#Transformations
cpm <- cpm(D_Raw)
lcpm <- cpm(D_Raw, log = TRUE)

#Density plot
nsamples <- ncol(D_Raw)
plot(density(lcpm[, 1]), ylim = c(0, 0.2), col = col.muscle[grep(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm))[1], names(col.muscle))], lwd = 2, las = 2, 
    main = "Filtered Data", xlab = "logCPM")
for(i in 2 : nsamples){
  den <- density(lcpm[, i])
  lines(den$x, den$y, lwd = 2, col = col.muscle[grep(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm))[i], names(col.muscle))])
}
```

### Corrected data for batch while muscle also considered in the model
```{r transFrilteringPlot3, echo = FALSE, warning = FALSE, fig.height = 6, fig.width = 10,, fig.show = 'hold'}
#Transformations
cpm <- cpm(D_BatchMus)
lcpm <- cpm(D_BatchMus, log = TRUE)
#Density plot
nsamples <- ncol(D_BatchMus)

par(mfrow = c(1, 2))
plot(density(lcpm[, 1]), ylim = c(0, 1.35), col = col.muscle[grep(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm))[1], names(col.muscle))], lwd = 2, las = 2, 
      main = "Raw Data\nBatch corrected(Muscle)", xlab = "logCPM")
for(i in 2 : nsamples){
  den <- density(lcpm[, i])
  lines(den$x, den$y, lwd = 2, col = col.muscle[grep(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm))[i], names(col.muscle))])
}

#Remove low expressed genes
keep.exprs <- filterByExpr(D_BatchMus, group = D_BatchMus$samples$group, min.count = 10)
D_BatchMus <- D_BatchMus[keep.exprs,, keep.lib.sizes = FALSE]

#Transformations
cpm <- cpm(D_BatchMus)
lcpm <- cpm(D_BatchMus, log = TRUE)

#Density plot
nsamples <- ncol(D_BatchMus)
plot(density(lcpm[, 1]), ylim = c(0, 0.2), col = col.muscle[grep(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm))[1], names(col.muscle))], lwd = 2, las = 2, 
    main = "Filtered Data", xlab = "logCPM")
for(i in 2 : nsamples){
  den <- density(lcpm[, i])
 lines(den$x, den$y, lwd = 2, col = col.muscle[grep(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm))[i], names(col.muscle))])
}

rm (list = setdiff(ls(), c("HTseqCount", "countAdjBatchMus", "D_Raw", "D_BatchMus", "withKit", "withoutKit", "col.muscle")))
```

#### **`Number of samples`**
```{r, echo = FALSE}
dim(D_Raw)[2]
```

#### **`Number of samples per muscle`**
```{r, echo = FALSE}
table(D_Raw$samples$group)
```

#### **`Number of muscles per individual`**
```{r, echo = FALSE}
count(D_Raw$samples, individual)
```

#### **`Number of genes`**
##### Raw
```{r, echo = FALSE}
dim(D_Raw)[1]
```

##### Batch corrected(including muscle in the model)
```{r, echo = FALSE}
dim(D_BatchMus)[1]
```

```{r colors, echo = FALSE, warning = FALSE, fig.height = 6, fig.width = 10}
#Set colors and labels
MusLabPlot <- gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(D_Raw))
MusColPlot <- col.muscle[match(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(D_Raw)), 
                                  names(col.muscle))]
IndLabPlot <- gsub("_.*", "", colnames(D_Raw))
col.individual <- c("MD06" = "#A6CEE3", "MD07" = "#1F78B4", "MD08" = "#B2DF8A", 
                     "MD10" = "#33A02C", "MD11" = "#FB9A99", "MD12" = "#E31A1C", 
                     "MD13" = "#FDBF6F", "MD14" = "#FF7F00", "MD15" = "#66C2A5", 
                     "MD16" = "#FC8D62", "MD17" = "#8DA0CB", "MD18" = "#E78AC3", 
                     "MD19" = "#A6D854", "MD20" = "#FFD92F", "MD21" = "#E5C494", 
                     "MD22" = "#B3B3B3", "MD26" = "#1B9E77", "MD27" = "#D95F02", 
                     "MD28" = "#7570B3", "MD31" = "#E7298A")
IndColPlot <- col.individual[match(gsub("_.*", "", colnames(D_Raw)), 
                                  names(col.individual))]
BatchLabPlot <- D_Raw$samples$batch[match(colnames(D_Raw), rownames(D_Raw$samples))]
col.batch <- c("1" = "#e6194B", "2" = "#3cb44b", "3" = "#4363d8", "4" = "#911eb4", "5" = "black")
BatchColPlot <- col.batch[match(D_Raw$samples$batch[match(colnames(D_Raw), 
                                                          rownames(D_Raw$samples))], 
                                names(col.batch))]
RNALabPlot <- D_Raw$samples$IsolationRNA[match(colnames(D_Raw), rownames(D_Raw$samples))]
col.RNA <- c("withKit" = "#999966", "withoutKit" = "#000000")
RNAColPlot <- col.RNA[match(D_Raw$samples$IsolationRNA[match(colnames(D_Raw), 
                                                             rownames(D_Raw$samples))], 
                                names(col.RNA))]
LaneLabPlot <- D_Raw$samples$lane[match(colnames(D_Raw), rownames(D_Raw$samples))]
col.lane <- c("L003.1" = "lightblue", "L004.1" = "darkblue", "L002.2" = "pink", "L004.2" = "red")
LaneColPlot <- col.lane[match(D_Raw$samples$lane[match(colnames(D_Raw), 
                                                       rownames(D_Raw$samples))], 
                                names(col.lane))]
HosLabPlot <- D_Raw$samples$Hospital[match(colnames(D_Raw), rownames(D_Raw$samples))]
col.Hos <- c("EMC" = "green", "HMC" = "orange")
HosColPlot <- col.Hos[match(D_Raw$samples$Hospital[match(colnames(D_Raw), 
                                                       rownames(D_Raw$samples))], 
                                names(col.Hos))]
```

## Normalization
```{r normalization, echo = FALSE, warning = FALSE, fig.height = 6, fig.width = 10}
#Normalization of gene expression distributions
D_Raw <- calcNormFactors(D_Raw, method = "TMM")
D_BatchMus <- calcNormFactors(D_BatchMus, method = "TMM")

#Transformations
cpm_Raw <- cpm(D_Raw)
lcpm_Raw <- cpm(D_Raw, log = TRUE)
cpm_BatchMus <- cpm(D_BatchMus)
lcpm_BatchMus <- cpm(D_BatchMus, log = TRUE)
```

#### **`Boxplot of logCPM values showing expression distributions for normalized data`**
- color = `muscle type`

```{r boxplot, fig.height = 6, fig.width = 16, echo = FALSE}
#Check distributions of samples using boxplots
par(mfrow = c(1, 1))
boxplot(lcpm_Raw, xlab = "", ylab = "logCPM", las = 2, main = "Boxplots of logCPM\nnormalised(raw)data", col = col.muscle[match(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm_Raw)), names(col.muscle))], names = gsub("_Sam.*", "", colnames(lcpm_Raw)), cex.axis = 0.5); abline(h = median(lcpm_Raw), col = "white", lty = 3)

boxplot(lcpm_BatchMus, xlab = "", ylab = "logCPM", las = 2, main = "Boxplots of logCPM\nnormalised(batch corrected-Muscle)data", col = col.muscle[match(gsub("MD[0-9][0-9]_|_Sam.*", "", colnames(lcpm_BatchMus)), names(col.muscle))], names = gsub("_Sam.*", "", colnames(lcpm_BatchMus)), cex.axis = 0.5); abline(h = median(lcpm_BatchMus), col = "white", lty = 3)
```

## Unsupervised clustering
### Aim of this step
Drawing exploratory plots to evaluate the similarities and dissimilarities between samples in an unsupervised manner.
 
### The input data
Normalized and transformed data(logCPM)

### 1. Principal component analysis(PCA)
We set center and scale. equal to TRUE in the call to prcomp to standardize the variables prior to the application of PCA(`prcomp(t(lcpm), center = TRUE, scale. = TRUE)`)
 
```{r PCA, echo = FALSE, warning = FALSE}
#Raw
PC_Raw <-  prcomp(t(lcpm_Raw), center = TRUE, scale. = TRUE)
PC_sd_Raw <- setNames(PC_Raw$sdev, paste0("PC", 1:length(PC_Raw$sdev)))#The standard deviations of the principal components
PC_var_expl_Raw <-(PC_sd_Raw^2)/ sum(PC_sd_Raw^2)* 100

#Batch corrected(muscle)
PC_BatchMus <-  prcomp(t(lcpm_BatchMus), center = TRUE, scale. = TRUE)
PC_sd_BatchMus <- setNames(PC_BatchMus$sdev, paste0("PC", 1:length(PC_BatchMus$sdev)))#The standard deviations of the principal components
PC_var_expl_BatchMus <-(PC_sd_BatchMus^2)/ sum(PC_sd_BatchMus^2)* 100

#Function to draw PC plots
plotPCA <- function(PCdata, PC_var_expl, PCx = "PC1", PCy = "PC2", 
                     col = NULL, main = NULL, labels = rownames(PCdata$x)){
  plot(PCdata$x[, PCx], PCdata$x[, PCy], cex = 1, col = col, 
        xlab = paste0(PCx, "(", round(PC_var_expl[PCx]), "%)"), 
        ylab = paste0(PCy, "(", round(PC_var_expl[PCy]), "%)"), 
        main = main, las = 1, )
  text(PCdata$x[, PCx], PCdata$x[, PCy], 
        labels = labels, cex = 0.6, pos = 2, col = col)
}
```

#### Scree plot
This plot shows the proportion of the variance in the data explained by 10 first principal components.
```{r screePlot, echo = FALSE, warning = FALSE, message = FALSE, fig.height = 6, fig.width = 10}
fviz_eig(PC_Raw, addlabels = TRUE, ylim = c(0, 18), main = "Raw")
fviz_eig(PC_BatchMus, addlabels = TRUE, ylim = c(0, 18), main = "Batch corrected including muscle")
```

#### PCA plots
#### Muscle type
```{r PCAPlots1, echo = FALSE, warning = FALSE, fig.height = 18, fig.width = 10, fig.show = 'hold'}
par(mfrow = c(2, 2))
plotPCA(PC_Raw, PC_var_expl_Raw, col = MusColPlot, main = "Muscle type\nNot corrected", labels = MusLabPlot)
plotPCA(PC_Raw, PC_var_expl_Raw, PCx = "PC3", PCy = "PC4", col = MusColPlot, labels = MusLabPlot)
plotPCA(PC_BatchMus, PC_var_expl_BatchMus, col = MusColPlot, main = "Muscle type\nBatch corrected(muscle)", labels = MusLabPlot)
plotPCA(PC_BatchMus, PC_var_expl_BatchMus, PCx = "PC3", PCy = "PC4", col = MusColPlot, labels = MusLabPlot)
```

#### Individual
```{r PCAPlots2, echo = FALSE, warning = FALSE, fig.height = 18, fig.width = 10, fig.show = 'hold'}
par(mfrow = c(2, 2))
plotPCA(PC_Raw, PC_var_expl_Raw, col = IndColPlot, main = "Individual\nNot corrected", labels = IndLabPlot)
plotPCA(PC_Raw, PC_var_expl_Raw, PCx = "PC3", PCy = "PC4", col = IndColPlot, labels = IndLabPlot)
plotPCA(PC_BatchMus, PC_var_expl_BatchMus, col = IndColPlot, main = "Individual\nBatch corrected(muscle)", labels = IndLabPlot)
plotPCA(PC_BatchMus, PC_var_expl_BatchMus, PCx = "PC3", PCy = "PC4", col = IndColPlot, labels = IndLabPlot)
```

#### Library preparation batch
```{r PCAPlots3, echo = FALSE, warning = FALSE, fig.height = 18, fig.width = 10}
par(mfrow = c(2, 2))
plotPCA(PC_Raw, PC_var_expl_Raw, col = BatchColPlot, main = "Batch\nNot corrected", labels = BatchLabPlot)
plotPCA(PC_Raw, PC_var_expl_Raw, PCx = "PC3", PCy = "PC4", col = BatchColPlot, labels = BatchLabPlot)

plotPCA(PC_BatchMus, PC_var_expl_BatchMus, col = BatchColPlot, main = "Batch\nBatch corrected(muscle)", labels = BatchLabPlot)
plotPCA(PC_BatchMus, PC_var_expl_BatchMus, PCx = "PC3", PCy = "PC4", col = BatchColPlot, labels = BatchLabPlot)

#Abbreviation for the paper
muscleAbbreviation <- c("GAL" = "GL", "GRA" = "GR", "VAL" = "VL", "VAM" = "VM", "SED" = "STD", "SEM" = "STM", "REF" = "RF")
names(col.muscle)<- muscleAbbreviation[match(names(col.muscle), names(muscleAbbreviation))]

(Tosave1 <- ggplot(PC_Raw$x[, 1:4] %>% as.data.frame() %>%
                   tibble::rownames_to_column(".id") %>%
                     mutate(Individual = gsub("_.*", "", .id)) %>%
                     mutate(Muscle = gsub("_.*", "", gsub("^[^_]*_", "", .id))) %>%
                     mutate(Batch = D_Raw$samples$batch[match(.id, row.names(D_Raw$samples))], 
                             Muscle = muscleAbbreviation[match(Muscle, names(muscleAbbreviation))]), 
                             aes(x = PC1, y = PC2, color = Batch, label = Muscle))+ 
  geom_point(size = 1)+
  scale_color_manual(values = col.batch)+
  theme_bw()+ 
    labs(fill = element_blank(), 
          x = paste0("PC1(", round(PC_var_expl_Raw["PC1"]), "%)"), 
          y = paste0("PC2(", round(PC_var_expl_Raw["PC2"]), "%)"))+
  theme(panel.grid = element_blank(), axis.text = element_text(size = 12), 
         legend.position = "none", axis.title = element_text(size = 12))+
    geom_text(aes(color = Batch), hjust = -0.1, vjust = 1, size = 3)+
      scale_x_continuous(n.breaks = 7))

(Tosave2 <- ggplot(PC_BatchMus$x[, 1:4] %>% as.data.frame() %>%
                   tibble::rownames_to_column(".id") %>%
                     mutate(Individual = gsub("_.*", "", .id)) %>%
                     mutate(Muscle = gsub("_.*", "", gsub("^[^_]*_", "", .id))) %>%
                     mutate(Batch = D_BatchMus$samples$batch[match(.id, row.names(D_BatchMus$samples))], 
                             Muscle = muscleAbbreviation[match(Muscle, names(muscleAbbreviation))]), 
                             aes(x = PC1, y = PC2, color = Batch, label = Muscle))+ 
  geom_point(size = 1)+
  scale_color_manual(values = col.batch)+
  theme_bw()+ 
    labs(fill = element_blank(), 
          x = paste0("PC1(", round(PC_var_expl_BatchMus["PC1"]), "%)"), 
          y = paste0("PC2(", round(PC_var_expl_BatchMus["PC2"]), "%)"))+
  theme(panel.grid = element_blank(), axis.text = element_text(size = 12), 
         legend.position = "none", axis.title = element_text(size = 12))+
    geom_text(aes(color = Batch), hjust = -0.1, vjust = 1, size = 3)+
      scale_x_continuous(n.breaks = 7))

ggsave(Tosave1, device = "tiff", units = "cm", width = 10, height = 10, filename = "figures/PCA_notCorrected.tif")
ggsave(Tosave2, device = "tiff", units = "cm", width = 10, height = 10, filename = "figures/PCA_Corrected.tif")

(Tosave3 <- ggplot(PC_BatchMus$x[, 1:4] %>% as.data.frame() %>%
                   tibble::rownames_to_column(".id") %>%
                     mutate(Individual = gsub("_.*", "", .id)) %>%
                     mutate(Muscle = gsub("_.*", "", gsub("^[^_]*_", "", .id))) %>%
                     mutate(Batch = D_BatchMus$samples$batch[match(.id, row.names(D_BatchMus$samples))], 
                             Muscle = muscleAbbreviation[match(Muscle, names(muscleAbbreviation))], 
                             Isolation = D_BatchMus$samples$IsolationRNA[match(.id, row.names(D_BatchMus$samples))]), 
                             aes(x = PC1, y = PC2, color = Isolation, label = Individual))+ 
  geom_point(size = 0.7)+
  scale_color_manual(values = col.RNA, name = "RNA isolation protocol", labels = c("miRNeasy Mini kit", "RNA precipitation"))+
  theme_bw()+ 
    labs(x = paste0("PC1(", round(PC_var_expl_BatchMus["PC1"]), "%)"), 
          y = paste0("PC2(", round(PC_var_expl_BatchMus["PC2"]), "%)"))+
  theme(panel.grid = element_blank(), axis.text = element_text(size = 12), 
         axis.title = element_text(size = 12))+
    geom_text(aes(color = Isolation), hjust = -0.1, vjust = 1, size = 1.5)+
      scale_x_continuous(n.breaks = 7))

ggsave(Tosave3, device = "tiff", units = "cm", width = 15, height = 10, filename = "figures/PCA_RNA.tif")

#For the figure below one should run the scripts from the beginning skipping lines 85-86
# (Tosave4 <- ggplot(PC_Raw$x[gsub("_Sam.*", "", rownames(PC_Raw$x)) %in%  gsub("_Sam.*", "", withKit), 1:4] %>% as.data.frame() %>%
#                    tibble::rownames_to_column(".id") %>%
#                      mutate(Individual = gsub("_.*", "", .id)) %>%
#                      mutate(Muscle = gsub("_.*", "", gsub("^[^_]*_", "", .id))) %>%
#                      mutate(Isolation = D_Raw$samples$IsolationRNA[match(.id, row.names(D_Raw$samples))]), 
#                              aes(x = PC1, y = PC2, color = Isolation, label = Individual))+ 
#   geom_point(size = 1)+
#   scale_color_manual(values = col.RNA, name = "RNA isolation protocol", labels = c("miRNeasy Mini kit", "RNA precipitation"))+
#   theme_bw()+ 
#         labs(x = paste0("PC1(32%)"), 
#           y = paste0("PC2(23%)"))+
#   theme(panel.grid = element_blank(), axis.text = element_text(size = 12), 
#          axis.title = element_text(size = 12))+
#     geom_text(aes(color = Isolation), hjust = 0.5, vjust = 1.1, size = 3)+
#       scale_x_continuous(n.breaks = 7))
# 
# ggsave(Tosave4, device = "tiff", units = "cm", width = 15, height = 10, filename = "PCA_RNA_5.tif")
```

#### PC association with different genes and factors
The contribution of each sample and gene in each PC is estimated.
The calculation step is explained by details within codes.
The gene contribution restult is not yet been used but the smaple contribution is used to show the association of different factors and PCs. To investigate the significance of association the Kruskal-Wallis test and Pearson correlation test are used for the categorical and numeric factors, respectively.
```{r PCAssociation, echo = FALSE, warning = FALSE, fig.height = 6, fig.width = 10}
#PCA results for genes(variables)
#Compute Coordinates
#The matrix of variable loadings(columns are eigenvectors)
for(file in c("Raw", "BatchMus")){
  PC <- get(paste0("PC_", file))
  loadings <- PC$rotation #Load of the genes in the PCs(g x n)
  sdev <- PC$sdev #Standard deviation for PCs(n)
  var_coord_func <- function(loadings, comp.sdev){loadings * comp.sdev}
  var.coord <- t(apply(loadings, 1, var_coord_func, sdev))#(g x n)
  #Compute Cos2
  var.cos2 <- var.coord ^ 2 #(g x n)
  #Compute contributions
  comp.cos2 <- apply(var.cos2, 2, sum) #Sum of the variance per PC
  contrib <- function(var.cos2, comp.cos2){var.cos2 * 100 / comp.cos2}
  var.contrib <- t(apply(var.cos2, 1, contrib, comp.cos2)) #%involvement of genes in variance of each PC
  #PCA results for the samples
  #Coordinates of samples
  Sam.coord <- PC$x
  #Cos2 of samples
  #1. square of the distance between an sample and the PCA center of gravity
  center <- PC$center #(g)the variable means(means that were substracted)
  scale <- PC$scale #(g)the variable standard deviations(the scaling applied to each variable)
  getdistance <- function(Sam_row, center, scale){return(sum(((Sam_row-center)/scale)^2))}
  lcpm <- get(paste0("lcpm_", file))
  d2 <- apply(t(lcpm), 1, getdistance, center, scale)
  #2. Compute the cos2. The sum of each row is 1
  cos2 <- function(Sam.coord, d2){return(Sam.coord^2/d2)}
  Sam.cos2 <- apply(Sam.coord, 2, cos2, d2)
  #Contributions of samples
  contrib <- function(Sam.coord, comp.sdev, n.Sam){100 *(1/n.Sam)* Sam.coord^2/comp.sdev^2}
  Sam.contrib <- t(apply(Sam.coord, 1, contrib, PC$sdev, nrow(Sam.coord)))
  #Add the sample information to the Sample contribution in the first 10 PCs
  D <- get(paste0("D_", file))
  Sam.contrib <- merge(Sam.contrib[, 1:10], D$samples, by =  "row.names")

  #Pearson correlation and Kruskall-Wallis tests
  GroupAsso <- matrix(data = NA, nrow = 10, ncol = 10)
  colnames(GroupAsso)<- c(colnames(Sam.contrib)[c(12, 16, 23, 20, 19, 21, 22, 18, 17, 13)])
  rownames(GroupAsso)<- colnames(Sam.contrib)[grepl("PC", colnames(Sam.contrib))]

  for(pc in  colnames(Sam.contrib)[grepl("PC", colnames(Sam.contrib))]){ 
    for(fac in colnames(Sam.contrib)[c(12, 16:20)]){ #Categorical factors
      fml <- as.formula(paste(noquote(pc), "~", fac))
      pval <- unlist(kruskal.test(fml, data = Sam.contrib)["p.value"])
      GroupAsso[pc, fac]<- round(pval, digits = 3)
    }
    for(fac in colnames(Sam.contrib)[c(13, 21:23)]){ #Numeric factors
      pval <- unlist(cor.test(Sam.contrib[, pc], Sam.contrib[, fac])["p.value"])
      GroupAsso[pc, fac]<- round(pval, digits = 3)
    }
  }
  assign(paste0("GroupAsso_", file), GroupAsso)
}
```

The below plot shows the association of biological and technical factors.
Each row corresponds to a PC and each column shows a factor. Cells contain the corresponding p-value(Kruskal-Wallis test or correlation test)for the association between a PC and the respective factor. 
The Kruskal-Wallis test and pearson correlation test are used for the categorical(black)and numeric(blue)factors, respectively. The table is color-coded by p-value according to the color legend.
```{r PCAssociationPlot, echo = FALSE, warning = FALSE, fig.height = 6, fig.width = 10}
colnames(GroupAsso)<- c("Muscle type", "Individual", "Age", "Hospital", "Isolation protocol", "RIN score", 
                           "Concentration", "Library batch", "Library size", "Sequencing lane")
fac.col <- c(rep("black", 2), "blue", rep("black", 2), "blue", "blue", "black", "blue", "black")
for(Asso in ls(pattern = "GroupAsso_")){
  GroupAsso <- get(Asso)
  GroupAssoPlot <- reshape2::melt(GroupAsso)
  print(ggplot(GroupAssoPlot, aes(x = Var2, y = Var1))+
    geom_tile(aes(fill = value))+
    scale_fill_gradientn(colours = c("white", "red", "black"), 
                          values = scales::rescale(c(0, 0.05, 0.1, 0.5, 1)))+
    geom_text(aes(label = format(value, digits = 2)))+ 
    ggtitle(paste0("PC associations with different factors\n", Asso))+ 
    labs(x = NULL, y = NULL, fill = "p-value")+ 
    theme(axis.text = element_text(size = 16), axis.title = element_text(size = 16), 
           axis.text.x = element_text(angle = 65, vjust = 1, hjust = 1, 
                                       color = fac.col)))
}
```

- For the not corrected data(Raw)the main factors contributing in the PC1 are batch, lane and RIN score.
- For the batch corrected data where the muscle was included in the model, only RIN was significantly contributed in the PC1.

### 2. Hierarchical clustering with heatmaps
Top 1000 most variable genes across all samples are used to draw a heatmap.
```{r hierarchicalHeatmapPlot, echo = FALSE, warning = FALSE, message = FALSE, fig.height = 10, fig.width = 14}
#Set colors
mypalette <- brewer.pal(11, "RdYlBu")
morecols <- colorRampPalette(mypalette)

myCols <- cbind(BatchColPlot, MusColPlot, IndColPlot)
colnames(myCols)<- c("Batch", "Muscle", "Individual")

for(file in c("Raw", "BatchMus")){
  lcpm <- get(paste0("lcpm_", file))
  #We estimate the variance for each gene(row)in the logCPM matrix
  countVar <- apply(lcpm, 1, var)
  #Get the row numbers for the top 1000 most variable genes
  highVar <- order(countVar, decreasing = TRUE)[1:1000]
  #Subset logCPM matrix
  hmDat <- lcpm[highVar, ]
  # Plot the heatmap
  library(heatmap3)
  heatmap3(hmDat, ColSideColors = myCols, 
                   labCol = gsub("_Sam.*", "", colnames(hmDat)), 
                   main = paste0("Top 1000 variable genes(", file, ")"), 
                   labRow = FALSE, scale = "row", 
                   col = rev(morecols(52)), breaks=seq(-5, 8, 0.25), margins = c(5, 1), 
            Rowv = T, showRowDendro = F)
}
save(D_BatchMus, countAdjBatchMus, file =  "outputs/correctedGenesDatasetToAnalysis.RData")
rm (list = setdiff(ls(), c("D_Raw", "D_BatchMus")))
```

- In the raw data, Samples from batch 5 are mainly clustered together. However, in general samples are mainly clustered by individual. These results suggest that the the individual effect is larger than the effect of other factors.
- In batch corrected data the clustering is mainly by individuals.

In the next step we perform analysis variance to have an estimation about the contribution of different factors on the expression gene level variations.

## Analysis of variance
### Aim of this analysis
Quantifying the relative contribution of technical and biological variation to the total variation and identifying the most important sources of technical variation.

### The input data
Normalized and transformed data: the mean-variance trend is removed(voom(limma)output).
Linear model is fitted for each gene considering muscle type, individual, isolation protocol, RIN score, concentration, library batch, sequencing lane, and library size as fixed effects.

```{r analysisOfVariance, echo = FALSE, warning = FALSE}
#Create a design matrix
design <- model.matrix(~0 + group + individual + 
                          batch, data = D_Raw$samples)
colnames(design)<- gsub("group|individual", "", colnames(design))

#Mean-variance trend was removed using the voom
for(Dfile in ls(pattern = "D_")){
  D <- get(Dfile)
  v <- voom(D, design)
  VarPart <- matrix(data = NA, nrow = nrow(v$E), ncol = 7)
  colnames(VarPart)<- c("Muscle type", "Individual", "RIN score", 
                         "Concentration", "Library batch", "Library size", 
                         "Residuals")
  rownames(VarPart)<- rownames(v$E)
  #Scale the numeric factors
  v$targets[, c(2, 10, 11)]<- scale(v$targets[, c(2, 10, 11)], center = T, scale = T)
  #Fit linear model for each gene(including all factors as fixed effects)
  for(gene in 1:nrow(v$E)){
    AnovaData <- cbind(v$E[gene, ], v$weights[gene, ], v$targets)
    colnames(AnovaData)[1:2]<- c("value", "weights")
    x <- anova(lm(value ~ 1 + group + individual + RIN + 
                    Concen. + batch + lib.size, weights = weights, data = AnovaData))
    VarPart[gene, ]<- 100*(x$`Sum Sq`)/sum((x$`Sum Sq`))
  }
  assign(gsub("D", "VarPart", Dfile), VarPart)
}
```

Below plot shows percentage variation in muscle gene expression dataset explained by different sources:

- Biological sources
    - Muscle type and Individuals
    
- Technical sources
    - RIN score, Concentration, Library preparation batch, and Library size

```{r analysisOfVariancePlot, echo = FALSE, warning = FALSE, message = FALSE, fig.height = 6, fig.width = 10}
for(VarPartfile in ls(pattern = "^VarPart_")){
  VarPart <- get(VarPartfile)
  colnames(VarPart)[1]<- "Muscle"
  VarPartPlot <- reshape2::melt(VarPart)
  print(Tosave <- ggboxplot(VarPartPlot, x = "Var2", y = "value", 
                    add.params = list(fill = "white"), outlier.shape = NA)+
           labs(x = element_blank(), y = "%Variance explained")+
           theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1, size = 12), 
                  axis.text.y = element_text(size = 12), axis.title =  element_text(size = 12), 
                  legend.position = "none")+ ylim(0, 100)+
      geom_boxplot(outlier.size = 0.2, color = c(rep("darkgreen", 2), rep("darkred", 4), "black")))

  assign(paste0("Tosave_", VarPartfile), Tosave)

}
ggsave(Tosave_VarPart_Raw, device = "tiff", units = "cm", width = 10, height = 10, filename = "figures/ANOVA_notCorrected.tif")
ggsave(Tosave_VarPart_BatchMus, device = "tiff", units = "cm", width = 10, height = 10, filename = "figures/ANOVA_Corrected.tif")
```

This plot shows that the biological factors have a high contribution in the variation. 
The individual factor shows the highest contribution. However, we should remember that other hidden technical factors could be also involved in the variation accounted for individual(the operation time/day, the differences in how fast samples were snap-freezed after the operation, ...).
The library preparation batch is the most important technical source of variation in the not corrected data but as expected its effect has been eliminated by batch effect correction.

### Print session info
```{r, warning = FALSE, message = FALSE}
sessionInfo()
```
