==> dataAnalyses/2.cellTypeComposition.Rmd <==
---
title: "Cell type analysis"
subtitle: 'Cell type composition'
output:
  html_notebook: default
  word_document: default
editor_options:
  chunk_output_type: console
  always_allow_html: true
---

'Note:' The muscle abbreviations were GRA(for Gracilis), SED(semitendinosus distal),  SEM(semitendinosus middle)), GAL(gastrocnemius lateralis), REF(rectus femoris), VAM(vastus medialis) and VAL(vastus lateralis) in all the analyses(scripts) and they were only changed to GR, STM, STD, GL, RF, VM and VL for the plots(shown in the publication).

```{r setup, include = FALSE}
#Empty the R environment
rm(list = ls())

#Set your working environment to the location where your current source file is saved into.
setwd(dirname(rstudioapi::getSourceEditorContext()$path))

#Check if libraries are already installed, otherwise install them
if(!"dplyr" %in% installed.packages())install.packages("dplyr")
if(!require("BiocManager", quietly = TRUE))install.packages("BiocManager")
if(!"limma" %in% installed.packages())BiocManager::install("limma")
if(!"edgeR" %in% installed.packages())BiocManager::install("edgeR")
if(!"ggplot2" %in% installed.packages())install.packages("ggplot2")
if(!"lsmeans" %in% installed.packages())install.packages("lsmeans")
if(!"RColorBrewer" %in% installed.packages())install.packages("RColorBrewer")
if(!"WGCNA" %in% installed.packages())install.packages("WGCNA")
if(!"reshape2" %in% installed.packages())install.packages("reshape2")

#Load installed libraries
suppressPackageStartupMessages({ 
  library(dplyr)
  library(limma)
  library(edgeR)
  library(ggplot2)
  library(lsmeans)
  library(RColorBrewer)
  library(WGCNA)
  })

#Load data
Data <- new.env(); load("outputs/correctedGenesDatasetToAnalysis.RData", envir = Data)
D_BatchMus <- Data$D_BatchMus
countAdjBatchMus <- Data$countAdjBatchMus

#Set colors
col.muscle <- c("GAL" = "#999999", "GRA" = "#0099FF", "REF" = "#FF9999", 
                "SED" = "#0033FF", "SEM" = "#0066FF", "VAL" = "#FF6666", 
                "VAM" = "#FF3333")
```

## The marker genes
Tendon gene markers are obtained from Kendal, et al., 2019, bioRxiv AND Orfei, et al., 2019, 
myofiber genes from Smith et al., 2013, 
Other cell type markers from Rubenstein et. al., 2020

```{r cellTypeMarkers, message = FALSE, warning = FALSE, echo = FALSE, fig.show = FALSE}
load("inputs/cellTypeMarkers.RData")
CellTypes <- unique(markers$`Cell-type`)

#Create a design matrix 
design <- model.matrix(~ group + individual, data = D_BatchMus$samples)
colnames(design) <- gsub("group|individual", "", colnames(design))
#Mean-variance trend was removed using the voom
v_BatchMus <- voom(D_BatchMus, design, plot = TRUE)

#Add the voom transformed data to the RData
save(D_BatchMus, countAdjBatchMus, v_BatchMus, file =  "outputs/correctedGenesDatasetToAnalysis.RData")

CellTypesEigengene <- matrix(data = NA, nrow = ncol(v_BatchMus), ncol = length(CellTypes))
rownames(CellTypesEigengene) <- colnames(v_BatchMus)
colnames(CellTypesEigengene) <- CellTypes

for(cell in CellTypes){
  GeneMarkers <- data.frame(hgnc_symbol = markers$hgnc_symbol.hg38[markers$`Cell-type`== cell]) %>%
    mutate(ensembl_gene_id = markers$ensembl_gene_id.hg38[markers$`Cell-type`== cell])
  Subset <- v_BatchMus$E[rownames(v_BatchMus$E) %in% GeneMarkers$ensembl_gene_id,]
  rownames(Subset) <- GeneMarkers$hgnc_symbol[match(rownames(Subset), GeneMarkers$ensembl_gene_id)]
  
  Eigengenes <- WGCNA::moduleEigengenes(t(Subset), rep("black", nrow(Subset)))$eigengenes
  
  if(all(rownames(CellTypesEigengene) == rownames(Eigengenes))){
    CellTypesEigengene[, cell] <- Eigengenes[, 1]}
  
  all(rownames(CellTypesEigengene[, cell]) == rownames(v_BatchMus$targets))
  CellData <- data.frame(CellTypesEigengene[, cell], v_BatchMus$targets)
  colnames(CellData)[1] <- "cell"
}

save(CellTypesEigengene, file = "outputs/cellTypeComposition/cellTypesGeneMarkersEigengene.RData")

CellTypeOverlappingGenes <- matrix(0, nrow = length(CellTypes), ncol = length(CellTypes))
colnames(CellTypeOverlappingGenes) <- rownames(CellTypeOverlappingGenes) <- CellTypes
COMBIN <- t(combn(CellTypes, 2))

for(com in 1:nrow(COMBIN)){
  CellTypeOverlappingGenes[COMBIN[com, 1], COMBIN[com, 2]] <- CellTypeOverlappingGenes[COMBIN[com, 2], COMBIN[com, 1]] <- length(intersect(markers$Gene[markers$`Cell-type` == COMBIN[com, 1]], markers$Gene[markers$`Cell-type` == COMBIN[com, 2]]))
}

colnames(CellTypeOverlappingGenes) <- rownames(CellTypeOverlappingGenes) <- paste0(colnames(CellTypeOverlappingGenes), "(", table(markers$`Cell-type`)[match(colnames(CellTypeOverlappingGenes), names(table(markers$`Cell-type`)))], " genes)")
print(ggplot(subset(reshape2::melt(CellTypeOverlappingGenes)[lower.tri(CellTypeOverlappingGenes),], 
                Var1 != Var2), 
        aes(x = Var1, y = Var2, fill = value)) +
  geom_tile(aes(fill = value), colour = "white") +
  geom_text(aes(label = value)) + 
  theme_classic() +
  scale_fill_gradient2(name="No. Genes", breaks = seq(0, 11, 2), labels = seq(0, 11, 2), 
                        low = "white", high = "red") + xlab(NULL) + ylab(NULL) +
  theme(axis.text.x = element_text(angle = 90, hjust =  1, vjust = 0.5), 
         axis.ticks = element_blank(), panel.border = element_blank(), 
         legend.position = "none") +
  scale_y_discrete(position = 'right'))
```

## Anova test on the eigengene values for each cell type
We have checked whether the differences are significant across muscles.

```{r cellTypeTest, message = FALSE, warning = FALSE, echo = FALSE, fig.height = 18, fig.width = 18}
CellTypesEigengeneToAnova <- merge(v_BatchMus$targets[, c("group", "individual")], CellTypesEigengene, by = "row.names")
colnames(CellTypesEigengeneToAnova) <- gsub(" |\\+", "", colnames(CellTypesEigengeneToAnova))

AnovaTest <- matrix(NA, nrow = length(colnames(CellTypesEigengeneToAnova)[-c(1:3)]), ncol = 43)
rownames(AnovaTest) <- colnames(CellTypesEigengeneToAnova)[-c(1:3)]
LSMEANS <- c()
for(cell in colnames(CellTypesEigengeneToAnova)[-c(1:3)]) {
  fml <- as.formula(paste(noquote(cell), "~ group +(1|individual)"))
  MM <- lmerTest::lmer(fml, data = CellTypesEigengeneToAnova)
  #Tukey and other multiple comparison tests can be performed with a handful of functions. The functions TukeyHSD, HSD.test, and LSD.test are probably not appropriate for cases where there are unbalanced data or unequal variances among levels of the factor, though TukeyHSD does make an adjustment for mildly unbalanced data.  It is my understanding that the multcomp and lsmeans packages are more appropriate for unbalanced data.  Another alternative is the DTK package that performs mean separation tests on data with unequal sample sizes and no assumption of equal variances.
  
  PairWiseComp <- as.data.frame(lsmeans::lsmeans(MM, pairwise ~ group)$contrasts)[c("contrast", "p.value", "t.ratio")] %>%
    arrange(contrast)
  AnovaTest[cell,] <- c(anova(MM)$`Pr(>F)`, PairWiseComp$p.value, PairWiseComp$t.ratio)
}

colnames(AnovaTest) <- c("Pr(>F)", PairWiseComp$contrast, paste("t.ratio:", PairWiseComp$contrast))
AnovaTest <- AnovaTest %>% as.data.frame(as.is = T)%>% tibble::rownames_to_column() %>% mutate(`FDR(>F)` =  p.adjust(`Pr(>F)`, method = "fdr")) %>% tibble::column_to_rownames("rowname")


AnovaTest$`t.ratio: REF - GRA` <- AnovaTest$`t.ratio: GRA - REF` * -1; AnovaTest$`t.ratio: GRA - REF` <- NULL
AnovaTest$`t.ratio: VAL - GRA` <- AnovaTest$`t.ratio: GRA - VAL` * -1; AnovaTest$`t.ratio: GRA - VAL` <- NULL
AnovaTest$`t.ratio: VAM - GRA` <- AnovaTest$`t.ratio: GRA - VAM` * -1; AnovaTest$`t.ratio: GRA - VAM` <- NULL
AnovaTest$`t.ratio: VAL - SED` <- AnovaTest$`t.ratio: SED - VAL` * -1; AnovaTest$`t.ratio: SED - VAL` <- NULL
AnovaTest$`t.ratio: VAM - SED` <- AnovaTest$`t.ratio: SED - VAM` * -1; AnovaTest$`t.ratio: SED - VAM` <- NULL
AnovaTest$`t.ratio: VAL - SEM` <- AnovaTest$`t.ratio: SEM - VAL` * -1; AnovaTest$`t.ratio: SEM - VAL` <- NULL
AnovaTest$`t.ratio: VAM - SEM` <- AnovaTest$`t.ratio: SEM - VAM` * -1; AnovaTest$`t.ratio: SEM - VAM` <- NULL


breaks <- seq(min(AnovaTest[, grep("t.ratio:", colnames(AnovaTest))]), 
                max(AnovaTest[, grep("t.ratio:", colnames(AnovaTest))]), length.out = 1000)
point1 <- which.min(abs(breaks - 3)) 
point2 <- which.min(abs(breaks + 3)) 
midpoint <- which.min(abs(breaks - 0))

rampCol1 <- colorRampPalette(c("darkblue", "blue"))(point2)
rampCol1.1 <- colorRampPalette(c("lightblue", "white"))(midpoint-point2)
rampCol2.1 <- colorRampPalette(c("white", "pink"))(point1-midpoint)
rampCol2 <- colorRampPalette(c("red", "darkred"))(length(breaks) - point1)
rampCols <- c(rampCol1, rampCol1.1, rampCol2.1, rampCol2)
rownames(AnovaTest) <- c("Endothelial Cells", "LUM+ FAP Cells", "Satellite Cells", "Pericytes", 
                           "FBN1+ FAP Cells", "NK Cells", "PCV Endothelial Cells", "T and B Cells", 
                           "Smooth Muscle Cells", "Myeloid Cells", "Fast-twitch", "Slow-twitch", "Tendon")
#Abbreviation for the plots
colnames(AnovaTest) <- gsub("VAL", " VL", colnames(AnovaTest))
colnames(AnovaTest) <- gsub("VAM", " VM", colnames(AnovaTest))
colnames(AnovaTest) <- gsub("REF", " RF", colnames(AnovaTest))
colnames(AnovaTest) <- gsub("GAL", " GL", colnames(AnovaTest))
colnames(AnovaTest) <- gsub("GRA", " GR", colnames(AnovaTest))
colnames(AnovaTest) <- gsub("SED", "STD", colnames(AnovaTest))
colnames(AnovaTest) <- gsub("SEM", "STM", colnames(AnovaTest))


pheatmap::pheatmap(t(AnovaTest[, grep("t.ratio:", colnames(AnovaTest))]), color = rampCols, scale = "none", legend_labels = "t.ratio", cluster_cols = T, cluster_rows = T, labels_row = gsub("-", "vs  ", gsub(".*:", "", grep("t.ratio: ", colnames(AnovaTest), value = T) )), fontsize = 12, angle_col = 90) 

CellTypeEVtoPlot <- CellTypesEigengene %>% 
  as.data.frame(as.is = T) %>%
  tibble::rownames_to_column("Row.names") %>%
  mutate(Muscle = gsub("_.*", "", gsub("^[^_]*_", "", Row.names))) %>% 
  dplyr::select(!c(Row.names)) %>%
  group_by(Muscle) %>%
  summarise_all(mean) %>%
  tibble::column_to_rownames("Muscle") 

breaks <- seq(min(CellTypeEVtoPlot), 
               max(CellTypeEVtoPlot), length.out = 100)
midpoint <- which.min(abs(breaks - 0))
rampCol1 <- colorRampPalette(c("blue", "white"))(midpoint)
rampCol2 <- colorRampPalette(c("white", "red"))(length(breaks)-(midpoint+1))
rampCols <- c(rampCol1, rampCol2)

#Abbreviation for the plots
rownames(CellTypeEVtoPlot) <- gsub("VAL", "VL", rownames(CellTypeEVtoPlot))
rownames(CellTypeEVtoPlot) <- gsub("VAM", "VM", rownames(CellTypeEVtoPlot))
rownames(CellTypeEVtoPlot) <- gsub("REF", "RF", rownames(CellTypeEVtoPlot))
rownames(CellTypeEVtoPlot) <- gsub("GAL", "GL", rownames(CellTypeEVtoPlot))
rownames(CellTypeEVtoPlot) <- gsub("GRA", "GR", rownames(CellTypeEVtoPlot))
rownames(CellTypeEVtoPlot) <- gsub("SED", "STD", rownames(CellTypeEVtoPlot))
rownames(CellTypeEVtoPlot) <- gsub("SEM", "STM", rownames(CellTypeEVtoPlot))
colnames(CellTypeEVtoPlot)[colnames(CellTypeEVtoPlot) == "Slow twitch"] = "Slow-twitch"
colnames(CellTypeEVtoPlot)[colnames(CellTypeEVtoPlot) == "Fast twitch"] = "Fast-twitch"

pheatmap::pheatmap(CellTypeEVtoPlot, 
                    color = rampCols, scale = "none", 
                    legend_labels = "t.ratio", 
                    cluster_cols = T, cluster_rows = T, 
                    fontsize = 12, key =F, angle_col = 90)
#Add hyphen 
colnames(CellTypesEigengene)[colnames(CellTypesEigengene) == "Slow twitch"] = "Slow-twitch"
colnames(CellTypesEigengene)[colnames(CellTypesEigengene) == "Fast twitch"] = "Fast-twitch"

CellTypesEigengeneToPlot <- reshape2::melt(CellTypesEigengene[, c("Endothelial Cells", "Fast-twitch", "Slow-twitch")])

CellTypesEigengeneToPlot$Muscle <- gsub("MD[0-9][0-9]_|_Sam.*", "", CellTypesEigengeneToPlot$Var1) 
ORDER <- c("GAL" = 7, "GRA" = 1, "REF" = 4, "SEM" = 2, "SED" = 3, "VAL" = 5, "VAM" = 6)
CellTypesEigengeneToPlot$order <- ORDER[match(CellTypesEigengeneToPlot$Muscle, names(ORDER))]

#Abbreviations for the plots
muscleAbbreviation <- c("GAL" = "GL", "GRA" = "GR", "VAL" = "VL", "VAM" = "VM", "SED" = "STD", "SEM" = "STM", "REF" = "RF")
CellTypesEigengeneToPlot$Muscle <- muscleAbbreviation[match(CellTypesEigengeneToPlot$Muscle, names(muscleAbbreviation))]
names(col.muscle) <- muscleAbbreviation[match(names(col.muscle), names(muscleAbbreviation))]

ggplot(CellTypesEigengeneToPlot, aes(x = forcats::fct_reorder(Muscle, order), y = value, fill = Muscle)) +
     geom_boxplot(alpha = .5) +
  scale_fill_manual(values =  col.muscle) + theme_bw() +
  facet_grid(~ Var2, scales = "free", space = "free_x") +
  theme(panel.grid = element_blank(), 
         axis.text.x = element_text(angle = 90, size = 12, vjust = 0.5, hjust = 1), axis.text.y = element_text(size = 12), 
         axis.ticks = element_blank(), legend.text = element_text(size = 12, vjust = 0.5), axis.title.y = element_text(size = 12), 
         strip.text = element_text(size = 12, face = "bold"), legend.position = "none") + 
  labs(y = "Eigenvalue of cell type specific gene markers", x = element_blank())

CellTypesEigengeneToPlot <- reshape2::melt(CellTypesEigengene[, !colnames(CellTypesEigengene) %in% c("Endothelial Cells", "Fast-twitch", "Slow-twitch")])

CellTypesEigengeneToPlot$Muscle <- gsub("MD[0-9][0-9]_|_Sam.*", "", CellTypesEigengeneToPlot$Var1) 
ORDER <- c("GAL" = 7, "GRA" = 1, "REF" = 4, "SEM" = 2, "SED" = 3, "VAL" = 5, "VAM" = 6)
CellTypesEigengeneToPlot$order <- ORDER[match(CellTypesEigengeneToPlot$Muscle, names(ORDER))]

#Abbreviations for the plots
CellTypesEigengeneToPlot$Muscle <- muscleAbbreviation[match(CellTypesEigengeneToPlot$Muscle, names(muscleAbbreviation))]

ggplot(CellTypesEigengeneToPlot, aes(x = forcats::fct_reorder(Muscle, order), y = value, fill = Muscle)) +
     geom_boxplot(alpha = .5) +
  scale_fill_manual(values =  col.muscle) + theme_bw() +
  facet_wrap(~ as.character(Var2), scales = "free", nrow = 4) +
  theme(panel.grid = element_blank(), 
         axis.text.x = element_text(angle = 90, size = 12, vjust = 0.5, hjust = 1), axis.text.y = element_text(size = 12), 
         axis.ticks = element_blank(), legend.text = element_text(size = 12, vjust = 0.5), axis.title.y = element_text(size = 12), 
         strip.text = element_text(size = 12, face = "bold"), legend.position = "none") + 
  labs(y = "Eigenvalue of cell type specific gene markers", x = element_blank())

```

## Gene and cell type association

```{r geneCellTypeCor, message = FALSE, warning = FALSE, echo = FALSE}
if(file.exists("outputs/cellTypeComposition/geneCellTypesEigengeneCor.RData"))
  {load("outputs/cellTypeComposition/geneCellTypesEigengeneCor.RData")} else {
    #Calculate the correlation and the p-value
    GeneCellCor <- matrix(NA, nrow = nrow(v_BatchMus$E), ncol = ncol(CellTypesEigengene))
    colnames(GeneCellCor) <- colnames(CellTypesEigengene)
    rownames(GeneCellCor) <- rownames(v_BatchMus$E)
    GeneCellCorPval <- GeneCellCor 
    for(gene in rownames(v_BatchMus$E)){
      for(cell in colnames(CellTypesEigengene)){
        if(all(names(v_BatchMus$E[gene,]) == names(CellTypesEigengene[, cell]))){
          GeneCellCor[gene, cell] <- cor(v_BatchMus$E[gene,], CellTypesEigengene[, cell])
          GeneCellCorPval[gene, cell] <- cor.test(v_BatchMus$E[gene,], CellTypesEigengene[, cell])$p.value
        }
      }
    }
    #Calculate the adjusted p-value
    GeneCellCorAdj <- GeneCellCorPval
    for(cell in colnames(GeneCellCorPval)){
      GeneCellCorAdj[, cell] <- round(p.adjust(GeneCellCorPval[, cell], method = "fdr"), 5)
    }
    GeneCellCorAdj <- as.data.frame(GeneCellCorAdj, as.is = T)
    GeneCellCor <- as.data.frame(GeneCellCor, as.is = T)
    
    #Subset the significant(FDR < 0.05) and moderate relationship(abs(cor) > 0.5)
    SigGenesCellTypes <- c()

    for(i in 1:ncol(GeneCellCorAdj)){
      AdjFilter <- rownames(GeneCellCorAdj[i] %>% filter(.[[1]] < 0.05))
      PosCorFilter <- rownames(GeneCellCor[i] %>% filter(.[[1]] >= 0.5))
      SigGenesCellTypes[[paste0(colnames(GeneCellCorAdj)[i], "(PosCor)")]] <- intersect(AdjFilter, PosCorFilter)
    }
    
    kable(sapply(SigGenesCellTypes, length)) 
    save(GeneCellCor, GeneCellCorPval, GeneCellCorAdj, SigGenesCellTypes, file = "outputs/cellTypeComposition/geneCellTypesEigengeneCor.RData")
  }
```

### Print session info
```{r, warning = FALSE, message = FALSE}
sessionInfo()
```
